Move core planning function to library
