+++ topological_action_planner/scripts/tap.py
@@ -22,6 +22,7 @@ import pykdl_ros
 from ed_py.utility import rooms_of_volume, rooms_of_entity
 from ed_py.world_model import WM
 from topological_action_planner.serialisation import from_dicts
+from topological_action_planner.planning import TopoPlanner
 from topological_action_planner.util import visualize, generate_dummy_graph
 from topological_action_planner.visualisation import create_tap_marker_array
 from cb_base_navigation_msgs.srv import GetPlan, GetPlanRequest, GetPlanResponse
@@ -29,35 +30,23 @@ from ed_navigation_msgs.srv import GetGoalConstraint
 from cb_base_navigation_msgs.msg import PositionConstraint
 
 
-def compute_path_length(path: List[PoseStamped]) -> float:
-    """
-    Calculate the total length of a path by summing up the lengths of all edges
-
-    :param path: List of PoseStampeds (assumed to be all in the same frame)
-    :return sum of cartesian distances between given PoseStamped
-    """
-    return sum(
-        math.hypot(a.pose.position.x - b.pose.position.x, a.pose.position.y - b.pose.position.y)
-        for a, b in zip(path, path[1:])
-    )
-
-
-class TopologicalActionPlanner:
+class TopoPlannerNode:
     def __init__(self):
         self.G = from_dicts(rospy.get_param("~edges"))
         self.plot = rospy.get_param("~plot", False)
 
         # TODO: These are magic numbers. Can be parametrized, but these are the defaults
-        self._action_costs = rospy.get_param(
+        action_costs = rospy.get_param(
             "~action_costs",
             {
                 Edge.ACTION_DRIVE: 1,  # per meter distance driven
                 Edge.ACTION_OPEN_DOOR: 5,  # per door opened
-                Edge.ACTION_PUSH_OBJECT: 10,
+                Edge.ACTION_PUSH_OBJECT: 10,  # per item pushed
             },
-        )  # per item pushed
+        )
 
-        self.wm = WM()
+        self._wm = WM()
+        self._topo_planner = TopoPlanner(self._wm, action_costs)
 
         self._get_constraint_srv = rospy.ServiceProxy("ed/navigation/get_constraint", GetGoalConstraint)
 
@@ -67,11 +56,9 @@ class TopologicalActionPlanner:
 
         # self.G = generate_dummy_graph()
         # to_yaml(self.G, '/tmp/graph.yaml')
-
-        self._global_planner = rospy.ServiceProxy("global_planner/get_plan_srv", GetPlan)
         if self.plot:
             visualize(self.G)
-        self._pub_grasp_marker.publish(create_tap_marker_array(self.G, self.wm))
+        self._pub_grasp_marker.publish(create_tap_marker_array(self.G, self._wm))
 
         rospy.loginfo("Topological action planner started")
 
@@ -83,9 +70,9 @@ class TopologicalActionPlanner:
         graph = copy.deepcopy(self.G)  # type: nx.Graph
 
         for node in graph.nodes.keys():
-            node0_entity = self.wm.get_entity(node[0])
+            node0_entity = self._wm.get_entity(node[0])
             graph.nodes[node]["room"] = rooms_of_volume(
-                self.wm, node0_entity, node[1]
+                self._wm, node0_entity, node[1]
             )  # Based on where they are in ED?
 
         if req.origin.entity == "":
@@ -93,10 +80,10 @@ class TopologicalActionPlanner:
             # We'll insert a node corresponding to our current position.
             # Difficulty is determining it's adjacency to other nodes, based on geometry
             robot_name = rospy.get_namespace().split("/")[-2]
-            robot_entity = self.wm.get_entity(robot_name)
+            robot_entity = self._wm.get_entity(robot_name)
             origin_node = robot_name, ""
             # Ask ED in which room the robot is currently, maybe based on it's pose
-            current_rooms = rooms_of_entity(self.wm, robot_entity)
+            current_rooms = rooms_of_entity(self._wm, robot_entity)
             current_room_ids = [r.uuid for r in current_rooms]
             # TODO: This assumes that the robot can always drive to any other node in the same room.
             # This might not always be true of course. It may also make more sense to only connect with the closest N waypoints
@@ -116,81 +103,9 @@ class TopologicalActionPlanner:
         rospy.loginfo("Requesting topological action plan from {} to {}".format(origin_node, dst_node))
 
         try:
-            lowest_total_cost = float("inf")
-            while True:
-                path = nx.shortest_path(graph, origin_node, dst_node, weight="weight")
+            edges = self._topo_planner.plan(graph, origin_node, dst_node)
 
-                edges = []
-                for u, v in zip(path, path[1:]):
-                    edges += [
-                        Edge(
-                            origin=Node(*u),
-                            destination=Node(*v),
-                            cost=graph[u][v]["weight"],
-                            action_type=graph[u][v]["action_type"],
-                        )
-                    ]
-
-                # For the drive edges, query the maybe now updated cost of driving that with current knowledge
-                for edge in edges:
-                    if edge.action_type == Edge.ACTION_DRIVE:
-                        # TODO: we should actually plan from the end of the plan found for the previous edge.
-                        # Otherwise the center pose could be blocked but not e whole area and we would still fail.
-                        entity = self.wm.get_entity(edge.origin.entity)
-                        if not entity:
-                            rospy.logwarn("No entity '{}'".format(edge.origin.entity))
-                            continue
-
-                        if edge.origin.area:
-                            src_vector = entity.volumes[edge.origin.area].center_point
-                            src = PoseStamped(
-                                header=Header(frame_id=entity.uuid),
-                                pose=Pose(
-                                    position=Point(src_vector.x(), src_vector.y(), 0),
-                                    orientation=Quaternion(0, 0, 0, 1),
-                                ),
-                            )
-                        else:  # In case we have eg. robot as the source
-                            src = tf2_ros.convert(entity.pose, PoseStamped)
-
-                        dst = self.get_area_constraint(edge.destination.entity, edge.destination.area)
-                        global_plan_res = self._global_planner(
-                            GetPlanRequest(start=src, goal_position_constraints=[dst])
-                        )
-
-                        if global_plan_res.succes:
-                            edge_cost = (
-                                compute_path_length(global_plan_res.plan) * self._action_costs[Edge.ACTION_DRIVE]
-                            )
-                            # edge_cost = 100 * 0.1 * self._action_costs[Edge.ACTION_DRIVE]
-                        else:
-                            # Cannot plan along this edge
-                            edge_cost = 100
-
-                        rospy.loginfo(
-                            "Updating cost of edge {} - {} = {}".format(
-                                edge.origin, edge.destination, edge_cost
-                            ).replace("\n", ", ")
-                        )
-                        graph[(edge.origin.entity, edge.origin.area)][(edge.destination.entity, edge.destination.area)][
-                            "weight"
-                        ] = edge_cost
-                        edge.cost = edge_cost
-
-                current_total_cost = sum([edge.cost for edge in edges])
-                if lowest_total_cost == current_total_cost:
-                    rospy.loginfo("The current plan has the lowest cost ({}), picking this".format(current_total_cost))
-                    break  # The cost is not changing anymore, we hit the optimum, use this path
-                elif current_total_cost < lowest_total_cost:
-                    rospy.loginfo(
-                        "The current plan has cost {}, lowest is {}. "
-                        "Trying a to find a better plan with updated edge costs".format(
-                            current_total_cost, lowest_total_cost
-                        )
-                    )
-                    lowest_total_cost = current_total_cost
-
-            self._pub_grasp_marker.publish(create_tap_marker_array(graph, self.wm))
+            self._pub_grasp_marker.publish(create_tap_marker_array(graph, self._wm))
             rospy.loginfo("Found plan of {} edges".format(len(edges)))
             return PlanResponse(error_msg="", error_code=PlanResponse.SUCCESS, edges=edges)
         except nx.NetworkXNoPath as no_path_found_ex:
@@ -225,18 +140,10 @@ class TopologicalActionPlanner:
         self.G[u][v]["weight"] = edge.cost
         self.G[u][v]["action_type"] = edge.action_type
 
-    def get_area_constraint(self, entity: str, area: str) -> PositionConstraint:
-        """What is the center pose of an entity and area?"""
-        res = self._get_constraint_srv(entity_ids=[entity], area_names=[area])
-        if not res.error_msg:
-            return PositionConstraint(constraint=res.position_constraint_map_frame, frame="map")
-        else:
-            raise Exception("Cannot get a constraint for {}.{}".format(entity, area))
-
 
 if __name__ == "__main__":
     rospy.init_node("topological_action_planner")
 
-    tap = TopologicalActionPlanner()
+    tap = TopoPlannerNode()
 
     rospy.spin()
